requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
sacrebleu>=2.3.1
unbabel-comet>=2.2.0
pandas>=2.0.0
scikit-learn>=1.3.0
sentencepiece>=0.1.99
accelerate>=0.21.0